services/mission-planning-agent/requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
pydantic==2.9.2
pyyaml==6.0.2
requests==2.32.3
python-dotenv==1.0.1
pyjwt==2.9.0
